Move {GenTx,Init}Cmd into gaia's new init package
